LayoutTests/http/tests/security/cross-frame-access-object-getPrototypeOf-in-put-expected.txt

This tests that you can't get the prototype of the window during [[Put]] operation.

PASS array.cocoa = cocoaValue threw exception SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match..
PASS 'str'.cocoa = cocoaValue threw exception SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match..
PASS: successfullyParsed should be 'true' and is.

TEST COMPLETE
